Extract storage-regen helpers to storage_regen.py
Move the bodies of _schedule_storage_regenerate, _spawn_only_generate,
_regen_already_failed_recently_async, _stamp_regen_failure, and
_finalize_regen_success (plus the _REGEN_FAILURE_TTL_SECONDS constant)
into a sibling module. The controller keeps thin bound-method
delegates so the regen state (_regenerate_pending / _regenerate_failed
/ _regenerate_lock) stays controller-owned and shared with monitor
callbacks; tests that patch the bound methods or instance attrs still
intercept.

Test patches that target controller-module symbols
(create_subprocess_exec, time.time) inside the regen body are
re-pointed at storage_regen.

Pure refactor; behaviour preserved. Drops ~210 lines from
controllers/devices/controller.py.
